+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.3.0
 * Handle multiple inputs
+* Resize advanced window when images are displayed
+* Add eye laterality classification
 
 ## 1.2.0
 * Take into account *NumberOfWorkers* params. By default, set at 0.5 : take half of the available workers